fix(web): issues with prerendering by moving auth.ts to root
