+++ apps/web/src/routes/(auth)/signin/+page.server.ts
@@ -1,4 +1,4 @@
-import { signIn } from '$lib/auth.js';
+import { signIn } from '../../../auth.js';
 import type { Actions } from './$types';
 
 export const actions = { default: signIn } satisfies Actions;
