+++ apps/web/src/routes/(auth)/signout/+page.server.ts
@@ -1,4 +1,4 @@
-import { signOut } from '$lib/auth.js';
+import { signOut } from '../../../auth.js';
 import type { Actions } from './$types';
 
 export const actions = { default: signOut } satisfies Actions;
